src/test/java/net/projecteuler/dustinbarnes/Problem22Test.java
package net.projecteuler.dustinbarnes;

import java.math.BigInteger;

import org.junit.Test;

import static org.hamcrest.CoreMatchers.equalTo;
import static org.junit.Assert.assertThat;


/**
 * http://projecteuler.net/index.php?section=problems&id=22
 *
 * Using names.txt (right click and 'Save Link/Target As...'), a 46K text file containing over
 * five-thousand first names, begin by sorting it into alphabetical order. Then working out the
 *  alphabetical value for each name, multiply this value by its alphabetical position in the
 * list to obtain a name score.
 *
 * For example, when the list is sorted into alphabetical order, COLIN, which is worth
 * 3 + 15 + 12 + 9 + 14 = 53, is the 938th name in the list. So, COLIN would obtain a score of 938  53 = 49714.
 *
 * What is the total of all the name scores in the file?
 */
public class Problem22Test
{
    Problem22 problem22 = new Problem22();

    @Test
    public void testSolve() throws Exception
    {
        assertThat(problem22.solve(), equalTo(new BigInteger("871198282")));
    }
}
